*All GPU resources are now released on exit **The Window class waits for each GPU to finish their work before releasing resources, which prevents the program from crashing on shutdown *Added some function comments in the Renderer class *Changed some, not all, member variables to the superior standard

+++ Test/D3D12/D3D12Renderer.hpp
@@ -77,7 +77,7 @@ private:
 	void SetUpRenderInstructions();
 	void ResetCommandListAndAllocator(int backbufferIndex, int index);
 	void MapMatrixData(int backBufferIndex);
-	void RecordRenderInstructions(D3D12Window* w, D3D12Camera* c, int commandListIndex, int backBufferIndex, int firstInstructionIndex, int numInstructions);
+	void RecordRenderInstructions(D3D12Window* w, D3D12Camera* c, int commandListIndex, int backBufferIndex, size_t firstInstructionIndex, size_t numInstructions);
 
 	void RecordCommands(int threadIndex);
 	void SetThreadWork(int threadIndex, D3D12Window* w, D3D12Camera* c, int backBufferIndex, int firstInstructionIndex, int numInstructions);
